ask-your-pdf/requirements.txt
streamlit==1.22.0
tiktoken==0.3.3
python-dotenv==1.0.0
PyPDF2==3.0.1
openai==0.27.6
langchain==0.0.158
faiss-cpu==1.7.4